backend/requirements.txt
# User Interface
streamlit==1.42.0
requests>=2.32.5

# Web Framework
fastapi==0.129.0
uvicorn[standard]==0.41.0
python-multipart==0.0.22

# AI & RAG Orchestration (Loosened versions to resolve conflicts)
groq==0.18.0
langchain
langchain-community
langchain-huggingface

# Vector Database & Embeddings
sentence-transformers==3.0.1
faiss-cpu==1.13.2
pypdf==6.7.1
python-dotenv==1.2.1

# Evaluation
pytest
httpx
langchain-core>=0.1.0
pytest-asyncio